More restructuring.
