+++ grammars/sexp.cg
@@ -1,6 +1,5 @@
-atom = fish;
-fish = f(i|)sh;
-# ignore = \s|\n|\t|\r;
-# 
-E = atom | fish;
-# E0 = | E E0;
+atom = [a-z0-9]+;
+ignore = \s|\n|\t|\r;
+
+E = atom | "(" E0 ")";
+E0 = | E E0;
